// projects/shared-app/match/map/map.enum.ts
export enum MatchMapFriendlyName {
    FiringRange = "The Firing Range",
    // Battle Royale Maps
    KingsCanyon = "Kings Canyon",
    Olympus = "Olympus",
    WorldsEdge = "World's Edge",
    // Arena Maps
    Artillery = "Artillery",
    GoldenGardens = "Golden Gardens",
    Overflow = "Overflow",
    Oasis = "Oasis",
    PartyCrasher = "Party Crasher",
    PhaseRunner = "Phase Runner",
    SkullTown = "Skull Town",
    ThermalStation = "Thermal Station",
}

export enum MatchMapGenericId {
    FiringRange = "firingrange",
    // Battle Royale Maps
    KingsCanyon = "kingscanyon",
    Olympus = "olympus",
    WorldsEdge = "worldsedge",
    // Arena Maps
    Artillery = "artillery",
    GoldenGardens = "goldengardens",
    Overflow = "overflow",
    Oasis = "oasis",
    PartyCrasher = "partycrasher",
    PhaseRunner = "phaserunner",
    SkullTown = "skulltown",
    ThermalStation = "thermalstation",
}